+++ Sources/Core/Extensions/String+Extension.swift
@@ -7,16 +7,8 @@
 //
 
 import Foundation
-import CryptoKit
 
 extension String {
-    var md5: String {
-        let digest = Insecure.MD5.hash(data: data(using: .utf8) ?? Data())
-        return digest.map {
-            String(format: "%02hhx", $0)
-        }.joined()
-    }
-    
     var urlEncoded: String {
         return self.addingPercentEncoding(withAllowedCharacters: .urlHostAllowed) ?? self
     }